+++ src/main/java/com/bcs/ventas/dao/mappers/ClienteMapper.java
@@ -9,6 +9,6 @@ import java.util.Map;
 @Mapper
 public interface ClienteMapper extends GeneralMapper<Cliente> {
 
-    int getTotalElements(Map<String, Object> var1);
+    Long getTotalElements(Map<String, Object> var1);
 
 }
