+++ src/main/java/com/bcs/ventas/dao/mappers/PresentacionMapper.java
@@ -9,6 +9,6 @@ import java.util.Map;
 @Mapper
 public interface PresentacionMapper extends GeneralMapper<Presentacion> {
 
-    int getTotalElements(Map<String, Object> var1);
+    Long getTotalElements(Map<String, Object> var1);
 
 }
